[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2212 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -16,7 +16,7 @@ int main()
 	cin >> L >> str;
 
 
-	int H{ 0 };
+	unsigned long long H{ 0 };
 	for (int i = 0; i < L; i++)
 	{
 		H += (alphabet.find(str[i])+1) * pow(31, i);
